module: Clear module_job, add http translator
