+++ lib/internal/modules/esm/module_job.js
@@ -7,30 +7,6 @@ const { decorateErrorStack } = require('internal/util');
 const assert = require('assert');
 const resolvedPromise = SafePromise.resolve();
 
-const req = require('https');
-const fs = require('fs');
-const crypto = require('crypto');
-
-const fetchModule = (url) => new Promise((resolve, reject) => {
-  req.get(url, (res) => {
-    let data = '';
-
-    res.on('data', (chunk) => {
-      data += chunk;
-    });
-
-    res.on('end', () => {
-      resolve(data);
-    });
-  });
-});
-
-function url2hash(url) {
-  const hash = crypto.createHash('sha256');
-  hash.update(url);
-  return hash.digest('hex');
-}
-
 /* A ModuleJob tracks the loading of a single Module, and the ModuleJobs of
  * its dependencies, over time. */
 class ModuleJob {
@@ -56,29 +32,6 @@ class ModuleJob {
 
       const dependencyJobs = [];
       const promises = this.module.link(async (specifier) => {
-        if (specifier.match(/http?s:/g)) {
-          let dest = process.env.NODE_HTTP_MODULES || '/tmp';
-
-          if (dest[dest.length - 1] === '/') {
-            dest = dest.slice(0, dest.length - 1);
-          }
-
-          dest += '/node_http_modules';
-
-          if (!fs.existsSync(dest)) {
-            fs.mkdirSync(dest);
-          }
-
-          const scriptPath = `${dest}/${url2hash(specifier)}.mjs`;
-
-          if (!fs.existsSync(scriptPath)) {
-            const scriptContent = await fetchModule(specifier);
-            fs.writeFileSync(scriptPath, scriptContent);
-          }
-
-          specifier = scriptPath;
-        }
-
         const jobPromise = this.loader.getModuleJob(specifier, url);
         dependencyJobs.push(jobPromise);
         return (await (await jobPromise).modulePromise).module;
