+++ lib/internal/modules/esm/translators.js
@@ -11,6 +11,8 @@ const internalURLModule = require('internal/url');
 const createDynamicModule = require(
   'internal/modules/esm/create_dynamic_module');
 const fs = require('fs');
+const https = require('https');
+const crypto = require('crypto');
 const { _makeLong } = require('path');
 const { SafeMap } = require('internal/safe_globals');
 const { URL } = require('url');
@@ -100,4 +102,55 @@ translators.set('json', async (url) => {
       throw err;
     }
   });
+});
+
+// Strategy for loading from HTTP(s)
+translators.set('http', async (url) => {
+  debug(`Translating Remote Http Module ${url}`);
+
+  const fetchModule = (url) => new Promise((resolve, reject) => {
+    https.get(url, (res) => {
+      let data = '';
+
+      res.on('data', (chunk) => {
+        data += chunk;
+      });
+
+      res.on('end', () => {
+        resolve(data);
+      });
+    });
+  });
+
+  function url2hash(url) {
+    const hash = crypto.createHash('sha256');
+    hash.update(url);
+    return hash.digest('hex');
+  }
+
+  let dest = process.env.NODE_HTTP_MODULES || '/tmp';
+
+  if (dest[dest.length - 1] === '/') {
+    dest = dest.slice(0, dest.length - 1);
+  }
+
+  dest += '/node_http_modules';
+
+  if (!fs.existsSync(dest)) {
+    fs.mkdirSync(dest);
+  }
+
+  const scriptPath = `${dest}/${url2hash(url)}.mjs`;
+
+  if (!fs.existsSync(scriptPath)) {
+    const scriptContent = await fetchModule(url);
+    fs.writeFileSync(scriptPath, scriptContent);
+  }
+
+  const source = `${await readFileAsync(new URL(scriptPath))}`;
+
+  return {
+    module: new ModuleWrap(stripShebang(source), scriptPath),
+    reflect: undefined
+  };
 });
